requirements.txt
torch==1.13.1+cu117
torchvision ==0.14.3+cu117
PySide6
dlib==19.7.0
opencv-python==4.5.5.64
Pillow==8.4.0
scikit-learn==0.24.2
ninja==1.11.1.1
ffmpeg
numpy==1.23.4
face_alignment==1.3.5
imageio==2.19.3
imageio-ffmpeg==0.4.7
librosa==0.9.2 # 
numba
resampy==0.3.1
pydub==0.25.1 
scipy==1.10.1
kornia==0.6.8
tqdm
yacs==0.1.8
pyyaml  
joblib==1.1.0
scikit-image==0.19.3
basicsr==1.4.2
facexlib==0.3.0
gradio
gfpgan
av
safetensors
